requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-huggingface==1.2.2
langchain-text-splitters==1.1.2
langgraph==1.1.9
anthropic==0.96.0
python-dotenv==1.2.2
faiss-cpu==1.13.2
sentence-transformers==5.4.1
streamlit==1.56.0
torch==2.11.0
torchvision==0.26.0
transformers==5.6.1